on_filter gray layer is common to all scatter_state users, default "grey"
